requirements.txt
asttokens==3.0.0
backcall==0.2.0
blinker==1.8.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.2.0
Flask==3.0.3
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==1.24.4
openpyxl==3.1.5
packaging==24.2
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
pyzmq==26.2.1
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
